+++ src/Symfony/Component/Messenger/Bridge/Amqp/Tests/Transport/ConnectionTest.php
@@ -102,6 +102,42 @@ class ConnectionTest extends TestCase
         );
     }
 
+    /**
+     * @group legacy
+     * @expectedDeprecation Invalid option(s) "foo" passed to the AMQP Messenger transport. Passing invalid options is deprecated since Symfony 5.1.
+     */
+    public function testDeprecationIfInvalidOptionIsPassedWithDsn()
+    {
+        Connection::fromDsn('amqp://host?foo=bar');
+    }
+
+    /**
+     * @group legacy
+     * @expectedDeprecation Invalid option(s) "foo" passed to the AMQP Messenger transport. Passing invalid options is deprecated since Symfony 5.1.
+     */
+    public function testDeprecationIfInvalidOptionIsPassedAsArgument()
+    {
+        Connection::fromDsn('amqp://host', ['foo' => 'bar']);
+    }
+
+    /**
+     * @group legacy
+     * @expectedDeprecation Invalid queue option(s) "foo" passed to the AMQP Messenger transport. Passing invalid queue options is deprecated since Symfony 5.1.
+     */
+    public function testDeprecationIfInvalidQueueOptionIsPassed()
+    {
+        Connection::fromDsn('amqp://host', ['queues' => ['queueName' => ['foo' => 'bar']]]);
+    }
+
+    /**
+     * @group legacy
+     * @expectedDeprecation Invalid exchange option(s) "foo" passed to the AMQP Messenger transport. Passing invalid exchange options is deprecated since Symfony 5.1.
+     */
+    public function testDeprecationIfInvalidExchangeOptionIsPassed()
+    {
+        Connection::fromDsn('amqp://host', ['exchange' => ['foo' => 'bar']]);
+    }
+
     public function testSetsParametersOnTheQueueAndExchange()
     {
         $factory = new TestAmqpFactory(