+++ src/Symfony/Component/Messenger/Bridge/Redis/CHANGELOG.md
@@ -6,3 +6,4 @@ CHANGELOG
 
  * Introduced the Redis bridge.
  * Added TLS option in the DSN. Example: `redis://127.0.0.1?tls=1`
+ * Deprecated use of invalid options